+++ src/utils/messages.js
@@ -2,20 +2,20 @@ import colors from 'colors'
 
 const { log } = console
 
-export const CLISuccesMessage = (fullThemeName, exemplaryComponent, themeName) => {
+export const CLISuccesMessage = (fullThemeName, exemplaryComponent, themeName, vendor) => {
   return log(
     colors.yellow('\nInstallation completed successfully!'),
-    colors.blue(`\n\n✅ Theme created in ${colors.green(`app/design/frontend/Snowdog/${themeName}`)}`),
+    colors.blue(`\n\n✅ Theme created in ${colors.green(`app/design/frontend/${vendor}/${themeName}`)}`),
     colors.blue('\n👉 Go to'),
     colors.green('Admin Panel -> Content -> Design -> Configuration'),
     colors.blue('and choose your theme'),
     (`(${colors.green(fullThemeName)})`),
     colors.blue(`\n✨ Use ${colors.green('yarn dev')} in`),
-    colors.green(`app/design/frontend/Snowdog/${themeName}/Snowdog_Components`),
+    colors.green(`app/design/frontend/${vendor}/${themeName}/Snowdog_Components`),
     colors.blue('to run components in Fractal'),
-    colors.blue('\n🔎 Read'),
+    colors.blue('\n🔎 Visit'),
     colors.green('Alpaca Docs'),
-    colors.blue('to learn how to use Alpaca Theme'),
+    colors.blue('at https://magento2-alpaca-docs.vercel.app/'),
     exemplaryComponent ? colors.blue(`\n\nVariables ${colors.green('$color-primary')} and ${colors.green('$button-text-color')} have been changed as an example`) : '',
     exemplaryComponent ? colors.blue(`\nSee ${colors.green(`_${themeName}-variables.scss`)} and ${colors.green(`_${themeName}-button-variables.scss`)} to edit them`) : '',
     colors.yellow('\n\n2022 Snowdog || https://snow.dog || https://github.com/SnowdogApps \n')