+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## [Unreleased]
 
 ## [Released]
+## [1.0.0] - 26.05.2022
+- Add custom vendor name functionality (#2)
+- Add mising styles imports, rework addFilesFromDir method (#10)
+- Import child theme variables before default ones (#13)
+- Add Alpaca Docs info (#1)
 ## [1.0.0-rc.2] - 25.03.2022
 - Improvements and initial fixes